Added text mesh pro essentials. After finalize interact Action now I am working on the inventory to show items in your inventory and the item you can take from other objects like a chest, wardrobe and so on

+++ Assets/Scripts/Items/ItemsDefinition/ItemGenericDefinition.cs
@@ -11,5 +11,5 @@ public class ItemGenericDefinition : ScriptableObject
     public string Name;
     public string Description;
     public float BasePrice;
-    public int MaxStack;
+    public byte MaxStack;
 }
